created BagSystem class for cleaner code
