+++ src/rpg/game/Tutorial.java
@@ -13,6 +13,9 @@ import rpg.systems.TutorialCombatSystem;
 import rpg.items.Consumable;
 import rpg.game.GameState;
 
+//new BagSystem import for cleaner code
+import rpg.systems.BagSystem;
+
 public class Tutorial {
     private final Scanner scanner;
     private final Random rand;
@@ -138,7 +141,7 @@ public class Tutorial {
 
                     case "bag":
                         try {
-                            StatusSystem.showBag(state);
+                            BagSystem.showBag(state);
                         } catch (Exception e) {
                             TextEffect.typeWriter("Unable to display bag right now.", 40);
                             System.err.println("Status error -> " + e.getMessage());
